Created AlarmDetails popup component and split files
